+++ src/graphgpt/__init__.py
@@ -1,6 +1,15 @@
 """GraphGPT public API."""
 
-from graphgpt.api import compile_workflow, inspect_workflow, load_workflow, validate_workflow
+from graphgpt.api import (
+    compile_workflow,
+    inspect_workflow,
+    load_workflow,
+    render_ecosystem_bundle,
+    validate_workflow,
+    write_ecosystem_bundle,
+)
+from graphgpt.application.ecosystem import EcosystemArtifact, InvocationContract
+from graphgpt.application.ports import EcosystemRenderer
 from graphgpt.plugin import (
     PLUGIN_API_VERSION,
     PLUGIN_ENTRY_POINT_GROUP,
@@ -17,7 +26,10 @@ __all__ = [
     "PLUGIN_API_VERSION",
     "PLUGIN_ENTRY_POINT_GROUP",
     "BindingRegistry",
+    "EcosystemArtifact",
+    "EcosystemRenderer",
     "GraphGPTPlugin",
+    "InvocationContract",
     "PluginCapability",
     "PluginInspection",
     "PluginManifest",
@@ -25,7 +37,9 @@ __all__ = [
     "inspect_installed_plugins",
     "inspect_workflow",
     "load_workflow",
+    "render_ecosystem_bundle",
     "validate_plugin",
     "validate_workflow",
+    "write_ecosystem_bundle",
 ]
 __version__ = "0.8.0"